restructured files + continued ChatSystem styling
