+++ frontend/src/Configs/DBConfig/Database.jsx
@@ -1,0 +1,2 @@
+// DATABASE CONFIGURATIONS PAGE
+// --------------------------------
