--- frontend/src/Routes/Chat.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import Sidebar from '../UIcomponents/ChatSystem/Sidebar/Sidebar';
-
-function Chat() {
-  return (
-    <div>
-    <Sidebar />
-    </div>
-  )
-}
-
-export default Chat;
