+++ frontend/src/Routes/Chat/Chat.css
@@ -1,0 +1,3 @@
+.app{
+    display: flex;
+}
