+++ frontend/src/Routes/Chat/Chat.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Sidebar from '../../UIcomponents/ChatSystem/Sidebar/Sidebar';
+import '../Chat/Chat.css';
+
+function Chat() {
+  return (
+    <div className='app'>
+    <Sidebar />
+
+    {/* <Chat /> */}
+    </div>
+  )
+}
+
+export default Chat;
