+++ frontend/src/UIcomponents/ChatSystem/Sidebar/Sidebar.jsx
@@ -1,6 +1,10 @@
 import React from 'react';
 import SidebarChannel from '../SidebarChannel/SidebarChannel';
-import '../Sidebar/Sidebar.css';
+import './Sidebar.css';
+
+//Placeholder for user profile image
+import profileImg from '../../../Assets/Images/profileImg.jpg';
+
 
 // MaterialUI Icons
 import { ExpandMore } from '@mui/icons-material';
@@ -8,6 +12,10 @@ import AddIcon from '@mui/icons-material/Add';
 import SignalCellularAltIcon from '@mui/icons-material/SignalCellularAlt';
 import InfoOutlinedIcon from "@mui/icons-material/InfoOutlined";
 import CallIcon from "@mui/icons-material/Call";
+import { Avatar } from '@mui/material';
+import MicIcon from '@mui/icons-material/Mic';
+import HeadsetIcon from '@mui/icons-material/Headset';
+import SettingsIcon from '@mui/icons-material/Settings';
 
 
 function Sidebar() {
@@ -31,23 +39,35 @@ function Sidebar() {
                 <div className="sidebar__channelsList">
                     <SidebarChannel />
                     <SidebarChannel />
-                    <SidebarChannel />
-                    <SidebarChannel />
                 </div>
             </div>
 
             <div className="sidebar__voice">
-            <SignalCellularAltIcon className="sidebar__voiceIcon" fontSize="large" />
-            <div className="sidebar__voiceInfo">
-            <h3> Voice Connected </h3>
-            <p> Stream </p>
-            </div>
-            <div className="sidebar__voiceIcons">
-            <InfoOutlinedIcon/>
-            <CallIcon/>
+                <SignalCellularAltIcon className="sidebar__voiceIcon" fontSize="large" />
+                <div className="sidebar__voiceInfo">
+                    <h3> Voice Connected </h3>
+                    <p> Stream </p>
+                </div>
 
+                <div className="sidebar__voiceIcons">
+                    <InfoOutlinedIcon />
+                    <CallIcon />
+                </div>
             </div>
 
+            <div className="sidebar__profile">
+                <Avatar src={profileImg} alt='Profile Image'/>
+                <div className="sidebar__profileInfo">
+                    <h3>Username</h3>
+                    <p>#ID</p>
+                </div>
+
+                <div className="sidebar__profileIcons">
+                    <MicIcon />
+                    <HeadsetIcon />
+                    <SettingsIcon />
+
+                </div>
             </div>
         </div>
     );
